+++ tests/Functional/FunctionalTest.php
@@ -868,6 +868,23 @@ class FunctionalTest extends WebTestCase
         ], json_decode($this->getModel('EntityWithUuid')->toJson(), true));
     }
 
+    public function testEntityWithTranslatable(): void
+    {
+        self::assertEquals([
+            'schema' => 'EntityWithTranslatable',
+            'type' => 'object',
+            'required' => ['translatable', 'translatableMessage'],
+            'properties' => [
+                'translatable' => [
+                    'type' => 'string',
+                ],
+                'translatableMessage' => [
+                    'type' => 'string',
+                ],
+            ],
+        ], json_decode($this->getModel('EntityWithTranslatable')->toJson(), true));
+    }
+
     public function testEntityWithIgnoredProperty(): void
     {
         self::assertEquals([
@@ -1260,5 +1277,80 @@ class FunctionalTest extends WebTestCase
         ];
 
         self::assertEquals($expected, json_decode($this->getModel('RangeInteger')->toJson(), true));
+    }
+
+    public function testSecuredApi(): void
+    {
+        $securitySchemes = $this->getOpenApiDefinition('secured')->components->securitySchemes;
+        self::assertCount(2, $securitySchemes);
+
+        $basicAuthScheme = $securitySchemes[0];
+        self::assertInstanceOf(OAAnnotations\SecurityScheme::class, $basicAuthScheme);
+        self::assertSame([
+            'securityScheme' => 'basicAuth',
+            'type' => 'http',
+            'scheme' => 'basic',
+        ], json_decode($basicAuthScheme->toJson(), true));
+
+        $apiKeyAuthScheme = $securitySchemes[1];
+        self::assertInstanceOf(OAAnnotations\SecurityScheme::class, $apiKeyAuthScheme);
+        self::assertSame([
+            'securityScheme' => 'apiKeyAuth',
+            'type' => 'apiKey',
+            'description' => 'API Key Authentication',
+            'name' => 'X-API-Key',
+            'in' => 'header',
+        ], json_decode($apiKeyAuthScheme->toJson(), true));
+
+        $operation = $this->getOperation('/secured/article/{id}', 'get', 'secured');
+
+        self::assertSame([
+            [
+                'basicAuth' => [
+                    'ROLE_USER',
+                ],
+            ],
+            [
+                'apiKeyAuth' => [
+                    'ROLE_USER',
+                ],
+            ],
+        ], $operation->security);
+
+        $operation = $this->getOperation('/secured/article', 'post', 'secured');
+
+        self::assertSame([
+            [
+                'basicAuth' => [
+                    'ROLE_USER',
+                    'ROLE_ADMIN',
+                ],
+            ],
+            [
+                'apiKeyAuth' => [
+                    'ROLE_USER',
+                    'ROLE_ADMIN',
+                ],
+            ],
+        ], $operation->security);
+
+        $operation = $this->getOperation('/secured/article/{id}', 'patch', 'secured');
+
+        self::assertSame([
+            [
+                'basicAuth' => [
+                    'ROLE_USER',
+                    'ROLE_ADMIN',
+                    'ROLE_UPDATE_ARTICLE',
+                ],
+            ],
+            [
+                'apiKeyAuth' => [
+                    'ROLE_USER',
+                    'ROLE_ADMIN',
+                    'ROLE_UPDATE_ARTICLE',
+                ],
+            ],
+        ], $operation->security);
     }
 }